feat: separated erc2612 permit logic into a separate utility function
