+++ src/utils/index.ts
@@ -1,3 +1,3 @@
+export * from "./permits";
 export * from "./abi";
 export * from "./ens";
-export * from "./permits";
