+++ src/utils/permits/index.ts
@@ -1,0 +1,1 @@
+export * from "./signERC2612Permit";
